Feature #69 lotus swagger and addtional feat (#93)
* feat(#69): class-validator 추가

* feat(#69): test login 추가

* feat(#69): lotus 기능들 Auth 헤더 확인 및 login 시 git token 갱신

* feat(#69): history 기능 Auth 헤더 확인

* feat(#69): lotus api dto 및 데코레이터 추가

* feat(#69): test 결과 토큰 json화, lotus 상세 조회 시 private이면 user확인 단계 추가

* feat(#69): lotus, user dto 추가 및 swagger docs 추가

* feat(#69): /user/gists 및 /user/lotus auth에서 유저값 읽어오기 적용

* feat(#69): lotus tag 연결

* feat(#69): 태그 search 기능 연결

* fix(#69): user login을 get에서 post로 수정

* fix(#69): serachTag->searchTag 오타 수정

* fix(#69): lotus entity 컬럼 중 category->tags로 수정

+++ apps/backend/src/lotus/lotus.service.ts
@@ -1,13 +1,18 @@
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
+import { In } from 'typeorm';
+import { LotusCreateRequestDto } from './dto/lotus.createRequest.dto';
 import { LotusDetailDto } from './dto/lotus.detail.dto';
 import { LotusDto } from './dto/lotus.dto';
 import { LotusPublicDto } from './dto/lotus.public.dto';
 import { LotusResponseDto } from './dto/lotus.response.dto';
+import { LotusUpdateRequestDto } from './dto/lotus.updateRequest.dto';
 import { MessageDto } from './dto/message.dto';
 import { SimpleUserResponseDto } from './dto/simple.user.response.dto';
 import { Lotus } from './lotus.entity';
 import { LotusRepository } from './lotus.repository';
 import { GistService } from '@/gist/gist.service';
+import { Tag } from '@/tag/tag.entity';
+import { TagService } from '@/tag/tag.service';
 import { UserService } from '@/user/user.service';
 
 @Injectable()
@@ -15,71 +20,105 @@ export class LotusService {
   constructor(
     private readonly lotusRepository: LotusRepository,
     private readonly gistService: GistService,
-    private readonly userService: UserService
+    private readonly userService: UserService,
+    private readonly tagService: TagService
   ) {}
   async createLotus(
+    userId: string,
     gitToken: string,
-    title: string,
-    isPublic: boolean,
-    tag: string[],
-    gistUuid: string,
-    inpLanguage: string,
-    inpVersion: string
+    lotusInputData: LotusCreateRequestDto
   ): Promise<LotusResponseDto> {
-    const language = !inpLanguage ? 'NodeJs' : inpLanguage;
-    const version: string = !inpVersion ? 'v22.11.0' : inpVersion;
-    const commits = await this.gistService.getCommitsForAGist(gistUuid, 1, gitToken);
+    if (!lotusInputData.language) {
+      lotusInputData.language = 'JavaScript';
+    }
+    if (!lotusInputData.version) {
+      lotusInputData.version = 'NodeJs:v22.11.0';
+    }
+    const commits = await this.gistService.getCommitsForAGist(lotusInputData.gistUuid, 1, gitToken);
     if (commits.length < 1) {
       throw new HttpException('this gist repository has no commit.', HttpStatus.NOT_FOUND);
     }
     const currentCommitId = commits[0].commitId;
 
-    if (await this.checkAlreadyExist(gistUuid, currentCommitId)) {
+    if (await this.checkAlreadyExist(lotusInputData.gistUuid, currentCommitId)) {
       throw new HttpException('same commit Lotus already exist.', HttpStatus.CONFLICT);
     }
-    const gitUser = await this.gistService.getUserData(gitToken);
-    const userData = await this.userService.findOne(gitUser.id);
-    await this.saveLotus(new LotusDto(title, isPublic, gistUuid, currentCommitId, userData, language, version));
+    const userData = await this.userService.findOneByUserId(userId);
+    const tags: Tag[] = await Promise.all(
+      lotusInputData.tags.map((tag) => {
+        return this.tagService.getTag(tag);
+      })
+    );
+    await this.saveLotus(new LotusDto(currentCommitId, userData, lotusInputData, tags));
     const lotusData = await this.lotusRepository.findOne({
-      where: { gistRepositoryId: gistUuid, commitId: currentCommitId },
-      relations: ['category']
+      where: { gistRepositoryId: lotusInputData.gistUuid, commitId: currentCommitId },
+      relations: ['tags']
     });
 
     return LotusResponseDto.ofSpreadData(SimpleUserResponseDto.ofUserDto(userData), lotusData);
   }
 
-  async updateLotus(lotusId: string, title: string, tag: string[], isPublic: boolean): Promise<LotusResponseDto> {
-    const result = await this.lotusRepository.update({ lotusId }, { title, isPublic });
-    if (!result.affected) throw new HttpException('update fail', HttpStatus.BAD_REQUEST);
+  async updateLotus(
+    lotusId: string,
+    lotusUpdateRequestDto: LotusUpdateRequestDto,
+    userIdWhoWantToUpdate: string
+  ): Promise<LotusResponseDto> {
     const updateLotus = await this.lotusRepository.findOne({
       where: { lotusId },
       relations: ['user']
     });
+    if (updateLotus.user.userId !== userIdWhoWantToUpdate) {
+      throw new HttpException('this is not allowed req', HttpStatus.FORBIDDEN);
+    }
     if (!updateLotus) throw new HttpException('invalid lotusId', HttpStatus.NOT_FOUND);
+    const result = await this.lotusRepository.update(
+      { lotusId },
+      { title: lotusUpdateRequestDto.title, isPublic: lotusUpdateRequestDto.isPublic }
+    );
+    if (!result.affected) throw new HttpException('update fail', HttpStatus.BAD_REQUEST);
     return LotusResponseDto.ofSpreadData(SimpleUserResponseDto.ofUserDto(updateLotus.user), updateLotus);
   }
 
-  async deleteLotus(lotusId: string): Promise<MessageDto> {
+  async deleteLotus(lotusId: string, userIdWhoWantToDelete: string): Promise<MessageDto> {
+    const deleteLotus = await this.lotusRepository.findOne({
+      where: { lotusId },
+      relations: ['user']
+    });
+    if (deleteLotus.user.userId !== userIdWhoWantToDelete) {
+      throw new HttpException('this is not allowed req', HttpStatus.FORBIDDEN);
+    }
+
     const result = await this.lotusRepository.delete({ lotusId });
     if (!result.affected) throw new HttpException('no match data', HttpStatus.NOT_FOUND);
 
     return new MessageDto('ok');
   }
 
-  async getLotusFile(gitToken: string, lotusId: string): Promise<LotusDetailDto> {
+  async getLotusFile(userId: string, gitToken: string, lotusId: string): Promise<LotusDetailDto> {
     const lotusData = await this.lotusRepository.findOne({
       where: { lotusId },
-      relations: ['category', 'user']
+      relations: ['tags', 'user']
     });
+    if (!lotusData.isPublic && lotusData.user.userId !== userId) {
+      throw new HttpException("this user can't access that lotus", HttpStatus.NOT_ACCEPTABLE);
+    }
+
     const commitFiles = await this.gistService.getCommit(lotusData.gistRepositoryId, lotusData.commitId, gitToken);
 
     return LotusDetailDto.ofGistFileListDto(commitFiles, lotusData);
   }
 
   async getPublicLotus(page = 1, size = 10, search: string): Promise<LotusPublicDto> {
+    const tags = await this.tagService.searchTag(search);
+
     const lotusData = await this.lotusRepository.find({
-      where: { isPublic: true },
-      relations: ['category', 'user']
+      where: {
+        isPublic: true,
+        tags: {
+          tagId: In(tags)
+        }
+      },
+      relations: ['tags', 'user']
     });
 
     const totalNum = lotusData.length;
@@ -92,7 +131,7 @@ export class LotusService {
   async getUserLotus(userId: string, page = 1, size = 10) {
     const lotusData = await this.lotusRepository.find({
       where: { user: { userId } },
-      relations: ['category', 'user']
+      relations: ['tags', 'user']
     });
     const totalNum = lotusData.length;
     const firstIdx = size * (page - 1);
